Splitting nodes.test.ts
