+++ tests/bin/nodes/find.test.ts
@@ -1,0 +1,185 @@
+import type { Host, Port } from '@/network/types';
+import type { NodeId } from '@/nodes/types';
+import os from 'os';
+import path from 'path';
+import fs from 'fs';
+import Logger, { LogLevel, StreamHandler } from '@matrixai/logger';
+import PolykeyAgent from '@/PolykeyAgent';
+import * as nodesUtils from '@/nodes/utils';
+import * as testBinUtils from '../utils';
+import * as testUtils from '../../utils';
+
+jest.mock('@/keys/utils', () => ({
+  ...jest.requireActual('@/keys/utils'),
+  generateDeterministicKeyPair:
+    jest.requireActual('@/keys/utils').generateKeyPair,
+}));
+
+describe('find', () => {
+  const password = 'password';
+  const logger = new Logger('find test', LogLevel.WARN, [
+    new StreamHandler(),
+  ]);
+  let dataDir: string;
+  let nodePath: string;
+  let passwordFile: string;
+  let polykeyAgent: PolykeyAgent;
+  let remoteOnline: PolykeyAgent;
+  let remoteOffline: PolykeyAgent;
+
+  let keynodeId: NodeId;
+  let remoteOnlineNodeId: NodeId;
+  let remoteOfflineNodeId: NodeId;
+
+  let remoteOnlineHost: Host;
+  let remoteOnlinePort: Port;
+  let remoteOfflineHost: Host;
+  let remoteOfflinePort: Port;
+
+  // Helper functions
+  function genCommands(options: Array<string>) {
+    return ['nodes', ...options, '-np', nodePath];
+  }
+
+  beforeAll(async () => {
+    dataDir = await fs.promises.mkdtemp(
+      path.join(os.tmpdir(), 'polykey-test-'),
+    );
+    nodePath = path.join(dataDir, 'keynode');
+    passwordFile = path.join(dataDir, 'passwordFile');
+    await fs.promises.writeFile(passwordFile, 'password');
+    polykeyAgent = await PolykeyAgent.createPolykeyAgent({
+      password,
+      nodePath: nodePath,
+      logger: logger,
+    });
+    keynodeId = polykeyAgent.nodeManager.getNodeId();
+
+    // Setting up a remote keynode
+    remoteOnline = await testUtils.setupRemoteKeynode({
+      logger,
+    });
+    remoteOnlineNodeId = remoteOnline.nodeManager.getNodeId();
+    remoteOnlineHost = remoteOnline.revProxy.getIngressHost();
+    remoteOnlinePort = remoteOnline.revProxy.getIngressPort();
+    await testUtils.addRemoteDetails(polykeyAgent, remoteOnline);
+
+    // Setting up an offline remote keynode
+    remoteOffline = await testUtils.setupRemoteKeynode({
+      logger,
+    });
+    remoteOfflineNodeId = remoteOffline.nodeManager.getNodeId();
+    remoteOfflineHost = remoteOffline.revProxy.getIngressHost();
+    remoteOfflinePort = remoteOffline.revProxy.getIngressPort();
+    await testUtils.addRemoteDetails(polykeyAgent, remoteOffline);
+    await remoteOffline.stop();
+
+    // Authorize session
+    await testBinUtils.pkStdio(
+      ['agent', 'unlock', '-np', nodePath, '--password-file', passwordFile],
+      {},
+      nodePath,
+    );
+  }, global.polykeyStartupTimeout * 3);
+  afterAll(async () => {
+    await polykeyAgent.stop();
+    await polykeyAgent.destroy();
+    await testUtils.cleanupRemoteKeynode(remoteOnline);
+    await testUtils.cleanupRemoteKeynode(remoteOffline);
+    await fs.promises.rm(dataDir, {
+      force: true,
+      recursive: true,
+    });
+  });
+
+  test('find an online node', async () => {
+    const commands = genCommands(['find', remoteOnlineNodeId]);
+    const result = await testBinUtils.pkStdio(commands, {}, dataDir);
+    expect(result.exitCode).toBe(0);
+    expect(result.stdout).toContain('Found node at');
+    expect(result.stdout).toContain(remoteOnlineHost);
+    expect(result.stdout).toContain(remoteOnlinePort);
+
+    // Checking json format.
+    const commands2 = genCommands([
+      'find',
+      remoteOnlineNodeId,
+      '--format',
+      'json',
+    ]);
+    const result2 = await testBinUtils.pkStdio(commands2, {}, dataDir);
+    expect(result2.exitCode).toBe(0);
+    expect(result2.stdout).toContain('success');
+    expect(result2.stdout).toContain('true');
+    expect(result2.stdout).toContain('message');
+    expect(result2.stdout).toContain(
+      `Found node at ${remoteOnlineHost}:${remoteOnlinePort}`,
+    );
+    expect(result2.stdout).toContain('host');
+    expect(result2.stdout).toContain('port');
+    expect(result2.stdout).toContain('id');
+    expect(result2.stdout).toContain(remoteOnlineNodeId);
+  });
+  test('find an offline node', async () => {
+    const commands = genCommands(['find', remoteOfflineNodeId]);
+    const result = await testBinUtils.pkStdio(commands, {}, dataDir);
+    expect(result.exitCode).toBe(0);
+    expect(result.stdout).toContain('Found node at');
+    expect(result.stdout).toContain(remoteOfflineHost);
+    expect(result.stdout).toContain(remoteOfflinePort);
+
+    // Checking json format.
+    const commands2 = genCommands([
+      'find',
+      remoteOfflineNodeId,
+      '--format',
+      'json',
+    ]);
+    const result2 = await testBinUtils.pkStdio(commands2, {}, dataDir);
+    expect(result2.exitCode).toBe(0);
+    expect(result2.stdout).toContain('success');
+    expect(result2.stdout).toContain('true');
+    expect(result2.stdout).toContain('message');
+    expect(result2.stdout).toContain(
+      `Found node at ${remoteOfflineHost}:${remoteOfflinePort}`,
+    );
+    expect(result2.stdout).toContain('host');
+    expect(result2.stdout).toContain('port');
+    expect(result2.stdout).toContain('id');
+    expect(result2.stdout).toContain(remoteOfflineNodeId);
+  });
+  test(
+    'fail to find an unknown node',
+    async () => {
+      const unknownNodeId = nodesUtils.makeNodeId(
+        'vrcacp9vsb4ht25hds6s4lpp2abfaso0mptcfnh499n35vfcn2gkg',
+      );
+      const commands = genCommands(['find', unknownNodeId]);
+      const result = await testBinUtils.pkStdio(commands, {}, dataDir);
+      expect(result.exitCode).toBe(1);
+      expect(result.stdout).toContain(`Failed to find node ${unknownNodeId}`);
+
+      // Checking json format.
+      const commands2 = genCommands([
+        'find',
+        unknownNodeId,
+        '--format',
+        'json',
+      ]);
+      const result2 = await testBinUtils.pkStdio(commands2, {}, dataDir);
+      expect(result2.exitCode).toBe(1);
+      expect(result2.stdout).toContain(`message`);
+      expect(result2.stdout).toContain(
+        `Failed to find node ${unknownNodeId}`,
+      );
+      expect(result2.stdout).toContain('id');
+      expect(result2.stdout).toContain(unknownNodeId);
+      expect(result2.stdout).toContain('port');
+      expect(result2.stdout).toContain('0');
+      expect(result2.stdout).toContain('host');
+      expect(result2.stdout).toContain('success');
+      expect(result2.stdout).toContain('false');
+    },
+    global.failedConnectionTimeout * 2,
+  );
+});
